Extract shared src↔convex modules into @margin-call/shared (#320).
Deduplicate ABIs, addresses helpers, clients, NAV math, and grant policy so Next and Convex share one source of truth, and wire mint actions through the tested decide* helpers.
