+++ convex/lib/chain/addresses.ts
@@ -1,8 +1,8 @@
 /**
  * Contract addresses from Convex env (mirrors NEXT_PUBLIC_* in the Next app).
+ * Address parsing lives in `@margin-call/shared`.
  */
-
-const ADDRESS_RE = /^0x[a-fA-F0-9]{40}$/;
+import { requireEnvAddress } from "@margin-call/shared";
 
 export type ConvexContractAddresses = {
   mockUsd: `0x${string}`;
@@ -11,23 +11,24 @@ export type ConvexContractAddresses = {
   ripEngine: `0x${string}`;
 };
 
-function requireAddress(envKey: string): `0x${string}` {
-  const value = process.env[envKey]?.trim();
-  if (!value) {
-    throw new Error(`${envKey} is not set in Convex env`);
+function requireConvexAddress(envKey: string): `0x${string}` {
+  try {
+    return requireEnvAddress(envKey);
+  } catch (err) {
+    const message = err instanceof Error ? err.message : String(err);
+    if (message === `${envKey} is not set`) {
+      throw new Error(`${envKey} is not set in Convex env`);
+    }
+    throw err;
   }
-  if (!ADDRESS_RE.test(value)) {
-    throw new Error(`${envKey} must be a 0x-prefixed 20-byte address`);
-  }
-  return value as `0x${string}`;
 }
 
 /** Addresses required for Starter Grant + pool indexing. */
 export function requireIndexerAddresses(): ConvexContractAddresses {
   return {
-    mockUsd: requireAddress("MOCKUSD_ADDRESS"),
-    packCustody: requireAddress("PACKCUSTODY_ADDRESS"),
-    assetRegistry: requireAddress("ASSETREGISTRY_ADDRESS"),
-    ripEngine: requireAddress("RIPENGINE_ADDRESS"),
+    mockUsd: requireConvexAddress("MOCKUSD_ADDRESS"),
+    packCustody: requireConvexAddress("PACKCUSTODY_ADDRESS"),
+    assetRegistry: requireConvexAddress("ASSETREGISTRY_ADDRESS"),
+    ripEngine: requireConvexAddress("RIPENGINE_ADDRESS"),
   };
 }
