+++ convex/tsconfig.json
@@ -8,7 +8,11 @@
     "skipLibCheck": true,
     "allowSyntheticDefaultImports": true,
     "isolatedModules": true,
-    "noEmit": true
+    "noEmit": true,
+    "paths": {
+      "@margin-call/shared": ["../packages/shared/src/index.ts"],
+      "@margin-call/shared/*": ["../packages/shared/src/*"]
+    }
   },
   "include": ["./**/*.ts"],
   "exclude": ["node_modules", "./__tests__/**", "./**/*.test.ts"]
