+++ packages/shared/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@margin-call/shared",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts",
+    "./abis": "./src/abis.ts",
+    "./address": "./src/address.ts",
+    "./chain": "./src/chain.ts",
+    "./clients": "./src/clients.ts",
+    "./deploy-blocks": "./src/deploy-blocks.ts",
+    "./nav": "./src/nav.ts",
+    "./parse-private-key": "./src/parse-private-key.ts",
+    "./starter-grant": "./src/starter-grant.ts",
+    "./stock-tokens": "./src/stock-tokens.ts"
+  },
+  "dependencies": {
+    "viem": "^2.55.10"
+  }
+}
