+++ packages/shared/src/deploy-blocks.ts
@@ -1,0 +1,9 @@
+/**
+ * Deploy block numbers for log-cursor bootstraps (from
+ * contracts/deployments/robinhood-testnet.*.json).
+ */
+export const PACKCUSTODY_DEPLOY_BLOCK = 95_307_505;
+export const RIPENGINE_DEPLOY_BLOCK = 95_311_248;
+
+export const PACKCUSTODY_DEPLOY_BLOCK_BIGINT = BigInt(PACKCUSTODY_DEPLOY_BLOCK);
+export const RIPENGINE_DEPLOY_BLOCK_BIGINT = BigInt(RIPENGINE_DEPLOY_BLOCK);
