+++ pnpm-workspace.yaml
@@ -1,1 +1,2 @@
-packages: []
+packages:
+  - "packages/*"
